+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 This file is used to list changes made in each version of the qmk_cookbook cookbook.
 
-# 0.1.0
+## [0.2.0] - TBD
+
+### Added
+- `users` recipe for creating an ubuntu user.
+- `directories` recipe for creating a the necessary directories associated with a user. 
+- `firmware` recipe used to clone qmk_firmware from your github account or qmk direct. 
+- Integration test for packages installed by the `.qmk_install` script.
+- Latent macOS support. 
+
+## [0.1.0] - 2018-10-04
 
 Initial release.